+++ failfast-junit-engine-it/src/test/kotlin/failfast/junit/it/fixtures/PendingTestFixtureTest.kt
@@ -2,9 +2,9 @@ package failfast.junit.it.fixtures
 
 import failfast.describe
 
-object IgnoredTestFixtureTest {
+object PendingTestFixtureTest {
     const val ROOT_CONTEXT_NAME = "the root context"
     val context = describe(ROOT_CONTEXT_NAME) {
-        itWill("pending test")
+        pending("pending test")
     }
 }